+++ src/Components/Navbar.js
@@ -24,7 +24,7 @@ function Navbar() {
           </StyleNavLink>
         </li>
         <li>
-          <StyleNavLink to="/edit">
+          <StyleNavLink to="/create">
             <svg
               xmlns="http://www.w3.org/2000/svg"
               fill="none"
