+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-import AuthenticationContextProvider from "./Contexts/AuthenticationContenxt";
+import AuthenticationContextProvider from "./Contexts/AuthenticationContext";
 import NotesContextProvider from "./Contexts/NotesContext";
 import GlobalStyles from "./Components/GlobalStyles";
 import App from "./App";
